+++ src/fluidSolver/sphSolver/sphZIndexSortedUniformGrid.hpp
@@ -1,3 +1,4 @@
+//  Copyright 2016 Robert Zhou
 //
 //  sphZIndexSortedUniformGrid.hpp
 //  MFluidSolver
@@ -5,19 +6,24 @@
 #ifndef MFLUIDSOLVER_FLUIDSOLVER_SPHZINDEXSORTEDUNIFORMGRID_HPP_
 #define MFLUIDSOLVER_FLUIDSOLVER_SPHZINDEXSORTEDUNIFORMGRID_HPP_
 
+#include <vector>
+
 #include "sphIndexSortedUniformGrid.hpp"
 #include "fluidSolver/zCurve.hpp"
 
 // Grid type XZY (X elements together, then Z, then Y)
 class SPHZIndexSortedUniformGrid : public SPHIndexSortedUniformGrid {
-public:
-  SPHZIndexSortedUniformGrid(const glm::vec3 &minBounds, const glm::vec3 &maxBounds, float cellSize, std::vector<SPHParticle> *master);
+ public:
+  SPHZIndexSortedUniformGrid(
+    const glm::vec3 &minBounds, const glm::vec3 &maxBounds,
+    float cellSize, std::vector<SPHParticle> *master);
   ~SPHZIndexSortedUniformGrid();
 
-  virtual unsigned long getIndex(unsigned int x, unsigned int y, unsigned int z);
+  virtual uint32_t getIndex(
+    unsigned int x, unsigned int y, unsigned int z);
 
-private:
+ private:
   ZCurve zCurve;
 };
 
-#endif /* MFLUIDSOLVER_FLUIDSOLVER_SPHZINDEXSORTEDUNIFORMGRID_HPP_ */
+#endif  // MFLUIDSOLVER_FLUIDSOLVER_SPHZINDEXSORTEDUNIFORMGRID_HPP_
